move thinking logic into its own package (#10990)
move thinking logic into its own package
